add work with user token to this app on view controller and add jwt_token provider

+++ statistics/src/main/go/com.github/entity/views_entity.go
@@ -7,7 +7,7 @@ import (
 type View struct {
 	Id        uint             `gorm:"column:id"`
 	CreatedAt time.Time        `gorm:"column:created_at"`
-	AccountId int64            `gorm:"column:account_id"`
+	AccountId uint             `gorm:"column:account_id"`
 	Products  []*ViewedProduct `gorm:"foreignkey:PurchaseId"`
 }
